Revert "added decrypt support"
This reverts commit ecc414f741256d3b3f80ced9c0eb8ff48fec7433.

The reasoning for doing this is that it has several typos (missing comma on the Crypt require, for example), and it brings in a dependency that happens to require node-gyp. bafcar does not require encrypted zip support at present, but does need some other bugfixes, so I'm reverting this commit for now.